:bug:
Update Version.java

+++ library/src/main/java/com/github/javiersantos/appupdater/objects/Version.java
@@ -25,9 +25,9 @@ public class Version implements Comparable<Version> {
         String[] thatParts = that.get().split("\\.");
         int length = Math.max(thisParts.length, thatParts.length);
         for (int i = 0; i < length; i++) {
-            int thisPart = i < thisParts.length ?
+            long thisPart = i < thisParts.length ?
                     Long.parseLong(thisParts[i]) : 0;
-            int thatPart = i < thatParts.length ?
+            long thatPart = i < thatParts.length ?
                     Long.parseLong(thatParts[i]) : 0;
             if (thisPart < thatPart)
                 return -1;
